Moved savings and added color to icon on asset page
